[MEASUREMENT]: Changes amount from int to double

+++ src/main/java/idatt2106v231/backend/model/ItemRefrigerator.java
@@ -44,7 +44,7 @@ public class ItemRefrigerator {
      *
      * @param amount the amount to add or remove
      */
-    public void updateAmount(int amount){
+    public void updateAmount(double amount){
         this.amount += amount;
     }
